Decompose plane into flying and parked

+++ src/Coordinate.java
@@ -1,8 +1,8 @@
 public class Coordinate {
-    public int x;
-    public int y;
+    public double x;
+    public double y;
 
-    public Coordinate(int x, int y) {
+    public Coordinate(double x, double y) {
         this.x = x;
         this.y = y;
     }
